Generates bootstrapped single cell data

SignacBoot uses a Seurat object or an expression matrix and performs feature selection, normalization and bootstrapping to generate a training data set to be used for cell type or cluster classification.

SignacBoot(
  E,
  L,
  labels,
  size = 1000,
  impute = TRUE,
  spring.dir = NULL,
  logfc.threshold = 0.25,
  p.val.adj = 0.05,
  verbose = TRUE
)

Arguments

E
a gene (rows) by cell (column) matrix, sparse matrix or a Seurat object. Rows are HUGO symbols.

L
cell type categories for learning.

labels
cell type labels corresponding to the columns of E.

size
Number of bootstrapped samples for machine learning. Default is 1,000.

impute
if TRUE, performs imputation prior to bootstrapping (see KSoftImpute). Default is TRUE.

spring.dir
if using SPRING, directory to categorical_coloring_data.json. Default is NULL.

logfc.threshold
Cutoff for feature selection. Default is 0.25.

p.val.adj
Cutoff for feature selection. Default is 0.05.

verbose
if TRUE, code speaks. Default is TRUE.

Value

Training data set (data.frame) to be used for building new models=.

See also

Examples

if (FALSE) {
# load Seurat object from SignacFast example
P <- readRDS("pbmcs.rds")

# run feature selection + bootstrapping to generate 2,000 bootstrapped cells
x = P@meta.data$celltypes
R_learned = SignacBoot(P, L = c("B.naive", "B.memory"), labels = x)
}